ThreadPool Class MVP seemingly working!
This took a bit of research, and even though it was not that much code
getting this done was very rewarding. There are a few key structural
concepts related to synchronization obviously that are settled in my
mind right now. There are optimizations that I could probably find
in my current code to improve performance.

There is customizability for changing texture resolution and amount of
threads to test performance.

I am interesting in a few ideas from here, potentially increasing
performance first or adding more functionality to this renderer.