Added FPS logging function, cleaned up files
printFPS is for profiling and runs as a detached thread that uses an
atomic int variable in simpleTexture.cpp to calculate the FPS.

Created a new utils.hpp file which will hold general functions I make.

I also created a shell script (not in this commit) that uses perf and
https://github.com/brendangregg/FlameGraph tool to generate a flamegraph
from the stack traces of my program as an interactable .svg file.

I also cleaned up all the constexpr stuff in simpleTexture.cpp and
scene.hpp to actually make sense.

Importantly, I currently have max optimizations on just to stress test
the FPS, but of course when developing and debugging I will turn it off
as neccessary.

The next steps are probably going to be separating my structs into
different files and splitting the .hpps to .hs and .cpps.